Updated bird.conf to be passive for ODL Lithium
ODL Lithium does not accept incoming connections from bird as it
incorrectly rejects the Peer AS. As such Bird must be passive and ODL
must establish the peering. This leads to an issue where ODL doesn't
peer from it's speaker IP and instead uses the same address as Bird [1].
The config now listens for ODL on the same address as Bird's BGP
instance.

[1]: https://lists.opendaylight.org/pipermail/bgpcep-users/2015-June/000340.html